Refactor

+++ app/src/main/java/org/example/factory/IReviewFactory.java
@@ -3,5 +3,5 @@ package org.example.factory;
 import org.example.models.reviews.Review;
 
 public interface IReviewFactory<T> {
-  Review createReview(int rating, String comment, T associatedEntity, Integer... additionalRatings);
+  Review createReview(Double rating, String comment, T associatedEntity, Double... additionalRatings);
 }
